Add files via upload
Uploaded what I have for the final project and my UI ideas- not much yet, but my png images are linked in the drawable folder. Have yet to make it work, but we have a basic set up going- where the days would go ('item0' and 'item1') where the buttons would be placed, etc. Obviously, it still looks boring as all get out.